backend/requirements.txt
Flask==2.1.0
requests==2.31.0
lxml==4.9.3
beautifulsoup4==4.12.2
google-cloud-storage==2.10.0